// sdks/typescript/models/TextToImageParams.ts
/**
 * Livepeer AI Runner
 * An application to run AI pipelines
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class TextToImageParams {
    'modelId': string;
    'prompt': string;
    'height'?: number;
    'width'?: number;
    'guidanceScale'?: number;
    'negativePrompt'?: string;
    'safetyCheck'?: boolean;
    'seed'?: number;
    'numInferenceSteps'?: number;
    'numImagesPerPrompt'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "modelId",
            "baseName": "model_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "prompt",
            "baseName": "prompt",
            "type": "string",
            "format": ""
        },
        {
            "name": "height",
            "baseName": "height",
            "type": "number",
            "format": ""
        },
        {
            "name": "width",
            "baseName": "width",
            "type": "number",
            "format": ""
        },
        {
            "name": "guidanceScale",
            "baseName": "guidance_scale",
            "type": "number",
            "format": ""
        },
        {
            "name": "negativePrompt",
            "baseName": "negative_prompt",
            "type": "string",
            "format": ""
        },
        {
            "name": "safetyCheck",
            "baseName": "safety_check",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "seed",
            "baseName": "seed",
            "type": "number",
            "format": ""
        },
        {
            "name": "numInferenceSteps",
            "baseName": "num_inference_steps",
            "type": "number",
            "format": ""
        },
        {
            "name": "numImagesPerPrompt",
            "baseName": "num_images_per_prompt",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return TextToImageParams.attributeTypeMap;
    }

    public constructor() {
    }
}